IOS-2257 Fix focus missing state
Проблема была в том, что при пересоздании view, стейт паблишера терялся. Судя по всему поведение с икскодом поменялось. Перенес в StateObject.
